backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Background Tasks
celery==5.3.6
redis==5.0.1

# Scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.40.0
tenacity==8.2.3

# AI
anthropic==0.18.1

# Utilities
python-dotenv==1.0.1
